Move tests from 'tests' to 'test'
The 'test' directory is automatically picked up by setuptools when
building the source distribution 'sdist'. One might want to get the
tests from the Python package index for some reason. In this case
it makes sense to have them in the source distribution, on the other
hand the tests are intentionally still kept out of binary
distributions.